Move InputBatch and CachedRequestState into gpu_input_batch.py
